pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cyberpunk"
version = "0.1.0"
description = "CLI network intelligence tool powered by local LLM"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "ollama>=0.4",
    "pyyaml>=6.0",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-ollama>=0.2",
    "langgraph>=0.2",
    "langfuse>=3.0",
    "python-dotenv>=1.0",
    "mac-vendor-lookup>=0.1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
cyberpunk = "cyberpunk.cli:app"

[tool.setuptools.packages.find]
include = ["cyberpunk*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
